角色列表修改

+++ CompetitionManagementSystem/src/main/java/com/qiaose/competitionmanagementsystem/mapper/ClassTableMapper.java
@@ -3,13 +3,13 @@ package com.qiaose.competitionmanagementsystem.mapper;
 import com.qiaose.competitionmanagementsystem.entity.ClassTable;
 
 public interface ClassTableMapper {
-    int deleteByPrimaryKey(Integer classId);
+    int deleteByPrimaryKey(Long classId);
 
     int insert(ClassTable record);
 
     int insertSelective(ClassTable record);
 
-    ClassTable selectByPrimaryKey(Integer classId);
+    ClassTable selectByPrimaryKey(Long classId);
 
     int updateByPrimaryKeySelective(ClassTable record);
